requirements.txt
git+https://github.com/xzzy/webtemplate.git#egg=webtemplate_dbca
Django==3.2.15
django-confy==1.0.4
whitenoise==5.3.0
django-cron==0.5.0
git+https://github.com/dbca-wa/dbca-utils.git#egg=dbca-utils
gunicorn==19.9.0
gevent==21.12.0
git+https://github.com/dbca-wa/django-media-serv.git#egg=django_media_serv
django-crispy-forms==1.14.0
django-countries==7.2.1
django-reversion==3.0.0
djangorestframework==3.12.0
djangorestframework-gis==0.16
djangorestframework-csv==2.0.0
geojson==1.3.3
psycopg2-binary==2.8.5
drf-spectacular==0.24.2
django-filter==22.1
